template/vite/src/backend/api.py
import webview as wv
import sys
import os
from pyder import *

class API:
    def __init__(self):
        self.window = wv.active_window()
        self.appID = f"{pyder_domainSystem}.{pyder_projectID}"
    
    def getConfigPath(self):
        if sys.platform == "win32":
            configPath = os.path.join(os.getenv("APPDATA"), self.appID)
        elif sys.platform == "darwin":
            configPath = os.path.join(os.getenv("HOME"), "Library", "Application Support", self.appID)
        elif sys.platform == "linux":
            configPath = os.path.join(os.getenv("HOME"), ".config", self.appID)
        else:
            configPath = os.path.join(os.getenv("HOME"), ".config", self.appID)
        return configPath