template/vite/requirements.txt
pyinstaller
pywebview; sys_platform != "linux"
pywebview[gtk]; sys_platform == "linux"